refactor(ui): update appStore import path
