+++ src/ui/App.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react'
 import classNames from 'classnames'
 import { AppRouter } from '@/ui/component/router/appRouter'
 import { Sidebar } from '@/ui/component/sidebar/sidebar'
-import { useAppStore } from '@/ui/store/appStore'
+import { useAppStore } from '@/ui/store/index'
 import { Toolbar } from '@/ui/component/toolbar/toolbar'
 import { ErrorBoundary } from '@/ui/component/errorBoundary/errorBoundary'
 import { Notification } from '@/ui/component/notification/notification'
