+++ src/ui/component/sidebar/languageSwitcher/languageSwitcher.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames'
 import type { FC } from 'react'
 import { useState, useCallback, useRef } from 'react'
 import { isCurrentLanguage } from '@/ui/i18n/utils'
-import { useAppStore } from '@/ui/store/appStore'
+import { useAppStore } from '@/ui/store/index'
 import { changeLanguage } from 'i18next'
 import { useOnClickOutside } from '@/ui/hook/useOnClickOutside'
 import { useOnKeyboard } from '@/ui/hook/useOnKeyboard'
