+++ src/ui/component/sidebar/sidebar.tsx
@@ -6,7 +6,7 @@ import classnames from 'classnames'
 import { Button } from '@/ui/component/button/button'
 import { Okp4Logo } from '@/ui/component/logo/okp4Logo'
 import { Switch } from '@/ui/component/switch/switch'
-import { useAppStore } from '@/ui/store/appStore'
+import { useAppStore } from '@/ui/store/index'
 import type { IconName } from '@/ui/component/icon/icon'
 import { Icon } from '@/ui/component/icon/icon'
 import { useBreakpoint } from '@/ui/hook/useBreakpoint'
