+++ src/ui/page/dataverse/dataverse.tsx
@@ -4,8 +4,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import InfiniteScroll from 'react-infinite-scroll-component'
 import { useBreakpoint } from '@/ui/hook/useBreakpoint'
-import { useAppStore } from '@/ui/store/appStore'
-import { useDataverseStore } from '@/ui/store/index'
+import { useDataverseStore, useAppStore } from '@/ui/store/index'
 import { DataverseItemCard } from '@/ui/view/dataverse/component/dataverseItemCard/dataverseItemCard'
 import { Button } from '@/ui/component/button/button'
 import Chip from '@/ui/component/chip/chip'
