+++ src/ui/page/dataverse/zone/governance/governanceNavigation.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react'
 import type { FC } from 'react'
 import { NavLink, useNavigate } from 'react-router-dom'
 import classnames from 'classnames'
-import { useAppStore } from '@/ui/store/appStore'
+import { useAppStore } from '@/ui/store/index'
 import type { IconName } from '@/ui/component/icon/icon'
 import { Icon } from '@/ui/component/icon/icon'
 import { useBreakpoint } from '@/ui/hook/useBreakpoint'
