+++ src/ui/page/error/internalError/internalError.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react'
 import { useNavigate } from 'react-router-dom'
 import classNames from 'classnames'
 import { useTranslation } from 'react-i18next'
-import { useAppStore } from '@/ui/store/appStore'
+import { useAppStore } from '@/ui/store/index'
 import { Button } from '@/ui/component/button/button'
 import { routes } from '@/ui/routes'
 import '@/ui/page/error/i18n/index'
